# docs/swagger.yaml
definitions:
  domain.PsConfig:
    properties:
      createdAt:
        type: string
      managers:
        type: string
      ownerID:
        type: integer
      psEvent:
        type: integer
      psFeature:
        type: integer
      psFilter:
        type: integer
      psID:
        type: integer
      psMessage:
        type: integer
      psScene:
        type: string
      psStrategy:
        type: integer
      updateUser:
        type: integer
      updatedAt:
        type: string
    type: object
  domain.User:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      id:
        type: integer
      money:
        type: string
      nickName:
        type: string
      passwordDigest:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      userName:
        type: string
    type: object
  domain.UserDTO:
    properties:
      key:
        type: string
      nick_name:
        type: string
      password:
        type: string
      user_name:
        type: string
    type: object
  middleware.CommonResponse:
    properties:
      data: {}
      errCode:
        type: string
      errMessage:
        type: string
      success:
        type: boolean
      type:
        type: string
    type: object
  middleware.Response:
    properties:
      errCode:
        type: string
      errMessage:
        type: string
      success:
        type: boolean
      type:
        type: string
    type: object
info:
  contact: {}
paths:
  /configs/list:
    get:
      consumes:
      - application/json
      description: 分页获取配置列表
      parameters:
      - description: 页码
        in: path
        name: pageNo
        required: true
        type: integer
      - description: 每页数量
        in: path
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 配置列表
          schema:
            allOf:
            - $ref: '#/definitions/middleware.CommonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.PsConfig'
                  type: array
              type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/middleware.Response'
      tags:
      - config
  /user/queryTest/{id}:
    get:
      consumes:
      - application/json
      description: 获取用户信息接口
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.CommonResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.User'
              type: object
        "400":
          description: error
          schema:
            $ref: '#/definitions/middleware.Response'
      tags:
      - user
  /users/register:
    post:
      consumes:
      - application/json
      description: Register a new user with the provided details
      parameters:
      - description: User DTO
        in: body
        name: userDto
        required: true
        schema:
          $ref: '#/definitions/domain.UserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.CommonResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: Register a new user
      tags:
      - users
swagger: "2.0"